PVS-50: Quick reorg of the typedefs
