+++ gql/resolvers/categories/custom/types.js
@@ -1,60 +1,6 @@
 import { gql } from 'apollo-server'
 
-export const commonCategoryTypes = gql`
-  scalar Date
-  enum ErrorTypes {
-    duplicate
-    failure
-    invalid
-    notFound
-    unauthorized
-  }
-
-  type Errors {
-    type: ErrorTypes
-    message: String
-  }
-
-  type RecipeCategory {
-    _id: ID!
-    label: String
-    createdAt: Date
-    updatedAt: Date
-  }
-  union RecipeCategoryItem = RecipeCategory | Errors
-
-  type RecipeCategories {
-    RecipeCategories: [RecipeCategory]
-  }
-  union RecipeCategoryItems = RecipeCategories | Errors
-
-  type BulkRecipeCategories {
-    all: [RecipeCategory]
-    failed: [String]
-    succeeded: [String]
-  }
-  union BulkRecipeCategoryItems = BulkRecipeCategories | Errors
-
-  type Query {
-    getAllCommonCategories: RecipeCategoryItems
-    getCommonCategoryById(id: ID!): RecipeCategoryItem
-    getCommonCategoryByLabel(label: String!): RecipeCategoryItem
-  }
-
-  type Mutation {
-    createCommonCategory(label: String!): RecipeCategoryItems
-    addBulkCommonCategories(labels: [String!]!): BulkRecipeCategoryItems
-    editCommonCategoryById(id: ID!, newLabel: String!): RecipeCategoryItems
-    deleteCommonCategoryById(id: ID!): RecipeCategoryItems
-    deleteCommonCategoryByLabel(label: String!): RecipeCategoryItems
-    deleteBulkCommonCategoriesById(ids: [ID!]!): BulkRecipeCategoryItems
-    deleteBulkCommonCategoriesByLabel(
-      labels: [String!]!
-    ): BulkRecipeCategoryItems
-  }
-`
-
-export const customCategoryTypes = gql`
+const customCategoryTypes = gql`
   scalar Date
   enum ErrorTypes {
     duplicate
@@ -123,3 +69,5 @@ export const customCategoryTypes = gql`
     ): BulkRecipeCategoryItems
   }
 `
+
+export default customCategoryTypes
