+++ gql/resolvers/faqs/types.js
