+++ gql/resolvers/patchNotes/types.js
