+++ gql/resolvers/typeDefs.js
@@ -1,0 +1,13 @@
+import userTypes from './user/types.js'
+import faqTypes from './faqs/types.js'
+import patchNoteTypes from './patchNotes/types.js'
+import commonCategoryTypes from './categories/common/types.js'
+import customCategoryTypes from './categories/custom/types.js'
+
+export default [
+  userTypes,
+  faqTypes,
+  patchNoteTypes,
+  commonCategoryTypes,
+  customCategoryTypes
+]
